Sort out media logos
