+++ src/content/media-outlets/index.js
@@ -1,0 +1,31 @@
+import wired from '../../assets/images/logos/wired.svg';
+import unhcr from '../../assets/images/logos/unhcr.png';
+import refugeesDeeply from '../../assets/images/logos/refugees-deeply.jpeg';
+import bbc from '../../assets/images/logos/bbc.svg';
+
+const mediaOutlets = [
+  {
+    url: 'http://www.wired.co.uk/article/codeyourfuture-refugee-coding-school',
+    src: wired,
+    alt: 'Wired',
+  },
+  {
+    url:
+      'http://www.unhcr.org/news/stories/2017/1/586e420c7/volunteers-train-refugees-to-crack-into-london-tech-industry.html',
+    src: unhcr,
+    alt: 'UNHCR',
+  },
+  {
+    url:
+      'https://www.newsdeeply.com/refugees/articles/2016/10/19/welcome-to-londons-refugee-coding-school',
+    src: refugeesDeeply,
+    alt: 'Refugees Deeply',
+  },
+  {
+    url: 'http://www.bbc.co.uk/programmes/p04yzrrg',
+    src: bbc,
+    alt: 'BBC',
+  },
+];
+
+export default mediaOutlets;
